Timing + SD added
Timing mit Tasks angepasst, SD-Logging hinzugefügt

+++ ESP32_CC_WIFI_DualCore/tachometer.h
@@ -6,7 +6,7 @@ void readSpeed() {
 
   if (prePuls == 0 && puls1 == 1) {
     pulsStop = micros();
-    unsigned long temp = pulsStop - pulsStart;
+    uint32_t temp = pulsStop - pulsStart;
     pulsStart = micros();
     for (int i = 0; i < 8; ++i) {
       pulsTime[8 - i] = pulsTime[7 - i];
